src/agents/pyproject.toml
[project]
name = "agents"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.12",
    "logfire>=3.14.0",
    "pydantic>=2.11.3",
    "pydantic-ai>=0.0.55",
    "python-dotenv>=1.1.0",
]
